+++ view/adminhtml/web/vue/field-type/template-list.vue
@@ -1,8 +1,5 @@
 <template>
-    <div
-        v-if="isVisible"
-        class="admin__fieldset fieldset-wide"
-    >
+    <div class="admin__fieldset fieldset-wide">
         <auto-complete
             :label="label[templateType]"
             :description="description[templateType]"
@@ -11,12 +8,8 @@
             :options="options"
             :default-option-value="defaultOptionValue"
             :config="config"
+            :disabled="disabled"
         />
-    </div>
-    <div
-        v-else
-    >
-        {{ noCustomTemplatesMessage[templateType] }}
     </div>
 </template>
 <script>
@@ -53,10 +46,6 @@ define(['Vue', 'mage/translate'], function(Vue, $t) {
                 description: {
                   'node': $t('Selected template'),
                   'submenu': $t('Selected template')
-                },
-                noCustomTemplatesMessage: {
-                  'node': $t('There is no custom node template for this type of node.'),
-                  'submenu': $t('There is no custom submenu (children wrapper) template for this type of node.')
                 }
             }
         },
@@ -82,8 +71,8 @@ define(['Vue', 'mage/translate'], function(Vue, $t) {
                 }
                 return list;
             },
-            isVisible () {
-                return this.options.length > 1
+            disabled () {
+                return this.options.length <= 1
             }
         },
         template: template